Merkle trees: renaming
