+++ src/libmerklecpp/test/paths.cpp
@@ -3,7 +3,7 @@
 #include <chrono>
 #include <cstdlib>
 #include <iostream>
-#include <merkle++.h>
+#include <merklecpp.h>
 
 #define PRINT_HASH_SIZE 3
